move functions to -> scripts and reduce wait time to 2 mins
