+++ scripts/appointments_ft.py
