+++ scripts/appointments_op.py
